refactor: import streamModelCall from ai instead of ai/internal
Use experimental_streamModelCall and Experimental_ModelCallStreamPart
from 'ai' (public exports) instead of ai/internal.
